pyproject.toml
[tool.poetry]
name = "modularmotifs"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
tk = "^0.1.0"
pillow = "^11.1.0"
lark = "^1.2.2"
matplotlib = "^3.10.1"
tkinter-tooltip = "^3.1.2"



[tool.poetry.group.ai]
optional = true

[tool.poetry.group.ai.dependencies]
diffusers = {git = "https://github.com/huggingface/diffusers", rev = "89e4d62"}
matplotlib = "^3.10.0"
accelerate = "^1.3.0"
omegaconf = "^2.3.0"
einops = "^0.8.0"
transformers = "^4.48.1"
scipy = "^1.15.1"
tensorboard = "^2.18.0"
openai-clip = "^1.0.1"
xformers = "^0.0.29.post1"
opencv-python = "^4.11.0.86"
torch = "*"
torchvision = "*"
torchaudio = "*"
scikit-learn = "^1.6.1"
peft = "^0.14.0"



[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
mypy = "^1.15.0"
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
